+++ lib/traps/__init__.py
@@ -2,6 +2,11 @@ from .QTrap import QTrap
 from .QTrapGroup import QTrapGroup
 from .QTrapOverlay import QTrapOverlay
 from .QTrapMenu import QTrapMenu
+from .commands import (
+    AddTrapCommand, RemoveTrapCommand,
+    MoveCommand, RotateCommand, WheelCommand)
 
 
-__all__ = 'QTrap QTrapGroup QTrapOverlay QTrapMenu'.split()
+__all__ = ('QTrap QTrapGroup QTrapOverlay QTrapMenu '
+           'AddTrapCommand RemoveTrapCommand '
+           'MoveCommand RotateCommand WheelCommand').split()